% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roanda.R
\name{authorize}
\alias{authorize}
\title{Connect to oanda authentication server.}
\usage{
authorize(accountType = c("live", "practice"), accountID = .accountID,
  token = .accountID)
}
\arguments{
\item{accountType}{Account Type (live or practice)}

\item{accountID}{API Account ID}

\item{token}{API token}
}
\description{
Connect to oanda authentication server.
}

