% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roanda.R
\name{getInstruments}
\alias{getInstruments}
\title{Get all available instruments for trade.}
\usage{
getInstruments(accountType = .accountType, token = .token,
  accountID = .accountID)
}
\arguments{
\item{accountType}{Account Type (live or practice)}

\item{token}{API token}

\item{accountID}{API Account ID}
}
\description{
Retrieve all available forex instruments available to trade.
}

