% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{global.assign}
\alias{global.assign}
\title{Shortcut for global variable assignment with logging and optional ZMQ storage backend.}
\usage{
global.assign(x, value, environ = .GlobalEnv)
}
\arguments{
\item{x}{A variable name.}

\item{value}{value assigned to x.}

\item{environ}{the environment to use.}
}
\description{
Shortcut for global variable assignment with logging and optional ZMQ storage backend.
}

