% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{package.require}
\alias{package.require}
\title{Attempt to load the package, and if unsuccessful try to install.}
\usage{
package.require(pkg = ".", quiet = TRUE)
}
\arguments{
\item{pkg}{Package name}

\item{quiet}{if TRUE suppresses output from this function.}
}
\description{
Attempt to load the package, and if unsuccessful try to install.
}
\examples{
package.require('quantmod')
}

