% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roanda-package.r
\docType{package}
\name{Roanda}
\alias{Roanda}
\alias{Roanda-package}
\title{Roanda.}
\description{
Easily create contract and instrument objects Includes functions to make
requesting historical data from oanda easier defines stocks
from vector of ticker symbols, a csv file, or a data.frame (package includes
dataset)
}
\author{
puglisij <admin@meatwad.org>
}

