% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roanda-package.r, R/zzz.r
\docType{package}
\name{Roanda}
\alias{Roanda}
\alias{Roanda-package}
\title{Roanda.}
\description{
Easily create contract and instrument objects Includes functions to make
requesting historical data from oanda easier defines stocks
from vector of ticker symbols, a csv file, or a data.frame (package includes
dataset)

\pkg{Roanda} etc
\code{Roanda} provides an interface to the oanda api

R OANDA Package
}
\section{Package options}{


Roanda uses the following \code{\link{options}} to configure behaviour:

\itemize{
  \item \code{Roanda.path}: path to use for \code{\link{dev_mode}}

  \item \code{Roanda.name}: your name, used when signing draft
    emails.

  \item \code{Roanda.desc.author}: a string providing a default Authors@R
    string to be used in new \file{DESCRIPTION}s.  Should be a R code, and
    look like \code{"Hadley Wickham <h.wickham@gmail.com> [aut, cre]"}. See
    \code{\link[utils]{as.person}} for more details.

  \item \code{Roanda.desc.license}: a default license string to use for
    new packages.

  \item \code{Roanda.desc.suggests}: a character vector listing packages to
    to add to suggests by defaults for new packages.
  \item \code{Roanda.desc}: a named list listing any other
    extra options to add to \file{DESCRIPTION}

}
}
\author{
puglisij <admin@meatwad.org>
}

